Add files via upload
The objective of this internship is to implement a LEMP stack (Linux, Nginx,
MySQL, PHP) on your local machine to simulate a banking system that
demonstrates the race conditions problem. They will use threads to make
concurrent requests to the money withdrawal endpoint and observe how these
simultaneous requests can cause inconsistencies in the system due to race
conditions.